separated passes and pipelines
