+++ src/graphics/api/api.h
@@ -17,7 +17,8 @@
 #include "device.h"
 #include "swapchain.h"
 #include "shader.h"
-#include "renderpasses.h"
+#include "passes.h"
+#include "pipelines.h"
 #include "framebuffers.h"
 #include "command.h"
